Base Path From Customer Login to Order Menu
This is a base level program from the login pane to the order menu for sprint 2. I used the panes from the first upload I did. I tried commenting as I went, sorry for the poor formatting. I tried making tasks at the beginning of the methods that need to be tested and completed. I was thinking about making a github project file so we can just build on top of one project instead of sending files back and forth. Let me know if you guys have any questions about how I set the methods up, this is my first time really setting up a project I'm not sure if there are better ways to achieve what we are trying to get done.