Program Update
Hey guys here's my update. I've gotten all the way through ordering a pizza and creating a receipt. What I'm going to work on now is creating a receipt class to handle the receipt with the current instance of the class and having the manager be able to search for that order as well as the user. So here's what's next on the plate:
1) Creating a customer class that holds an array of customers.
       - We only need to show an example of a few customers so we can just make an object array of the class with a set capacity.
2) Creating a manager class that holds an array of managers.
3) Creating methods to check login information with array of customers/managers
All the variables needed for the new classes are all at the top of the program. Let me know if I'm missing anything else or whatever comes to mind.
Also I have been doing a poor job of commenting my code I'll try to clean it up.